requirements.txt
torch>=1.12.0
torchvision
torch-geometric>=2.2.0
pytorch-lightning>=2.0.0
scikit-learn
matplotlib
numpy
pandas
seaborn
wandb
PyYAML